requirements.txt
pandas
numpy
pandas-ta
yfinance
ccxt
feedparser
vaderSentiment
scikit-learn
plotly
APScheduler
python-dotenv
loguru
pyarrow
PyYAML
yfinance
ccxt

requests
beautifulsoup4
python-dateutil

